backend/requirements.txt
pymongo[srv]==4.6.1
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
